Fix entity reference error
When using an object that wraps the ID property, even though we typecast to a string inside the ID getter we get an exception on the second call for the same ID.

It is caused by not using the ID getter for the getById repository method.

This PR fixes it by internally typecasting everywhere the id to a string.